* Decent shadows * Started work on a separate scroller component that I hope will actually work properly

+++ client-gui-jfx/src/main/java/com/hypersocket/client/gui/jfx/Configuration.java
@@ -144,7 +144,7 @@ public class Configuration {
 		avoidReserved.addListener(new BooleanPreferenceUpdateChangeListener(
 				node, "avoidReserved"));
 
-		size.set(node.getInt("size", 48));
+		size.set((int)node.getDouble("size", 48));
 		size.addListener(new ChangeListener<Number>() {
 
 			@Override
